feat: add SolidityNode provider support
