+++ crates/provider/src/provider/solidity.rs
@@ -1,0 +1,339 @@
+//! Read-only provider over a TRON SolidityNode.
+
+use core::time::Duration;
+use std::sync::Arc;
+
+use tronz_primitives::{Address, TxId};
+
+use crate::{
+    Result,
+    error::ProviderError,
+    provider::pending::PendingTransactionError,
+    transport::{
+        SolidityTransport,
+        grpc::{RetryConfig, SolidityGrpcTransport, SolidityGrpcTransportBuilder},
+    },
+    types::{
+        AccountInfo, BlockInfo, ConstantCallResult, SignedTransaction, TransactionInfo,
+        TriggerSmartContract,
+    },
+};
+
+const DEFAULT_POLL_INTERVAL: Duration = Duration::from_secs(3);
+const DEFAULT_POLL_ATTEMPTS: u32 = 20;
+
+/// A read-only provider over `protocol.WalletSolidity`.
+#[derive(Clone)]
+pub struct SolidityProvider<T: SolidityTransport = SolidityGrpcTransport> {
+    inner: Arc<T>,
+}
+
+impl<T: SolidityTransport> SolidityProvider<T> {
+    /// Wrap an existing [`SolidityTransport`].
+    pub fn new(transport: T) -> Self {
+        Self { inner: Arc::new(transport) }
+    }
+
+    /// Borrow the underlying transport.
+    pub fn transport(&self) -> &T {
+        &self.inner
+    }
+
+    /// Fetch the latest solidified block.
+    pub async fn get_now_block(&self) -> Result<BlockInfo> {
+        self.inner.get_now_block().await.map_err(|e| ProviderError::from(e.into()))
+    }
+
+    /// Fetch a solidified block by height.
+    pub async fn get_block_by_number(&self, num: i64) -> Result<BlockInfo> {
+        self.inner.get_block_by_number(num).await.map_err(|e| ProviderError::from(e.into()))
+    }
+
+    /// Fetch solidified account state.
+    pub async fn get_account(&self, address: Address) -> Result<AccountInfo> {
+        self.inner.get_account(address).await.map_err(|e| ProviderError::from(e.into()))
+    }
+
+    /// Fetch a transaction by id from solidified state.
+    pub async fn get_transaction(&self, tx_id: TxId) -> Result<SignedTransaction> {
+        self.inner.get_transaction_by_id(tx_id).await.map_err(|e| ProviderError::from(e.into()))
+    }
+
+    /// Fetch a transaction's receipt from solidified state.
+    ///
+    /// Returns `None` until the transaction has solidified — this is the signal
+    /// [`wait_for_transaction`](Self::wait_for_transaction) polls on.
+    pub async fn get_transaction_info(&self, tx_id: TxId) -> Result<Option<TransactionInfo>> {
+        self.inner.get_transaction_info(tx_id).await.map_err(|e| ProviderError::from(e.into()))
+    }
+
+    /// Fetch all transaction receipts in a solidified block.
+    pub async fn get_transaction_info_by_block_num(
+        &self,
+        block_num: i64,
+    ) -> Result<Vec<TransactionInfo>> {
+        self.inner
+            .get_transaction_info_by_block_num(block_num)
+            .await
+            .map_err(|e| ProviderError::from(e.into()))
+    }
+
+    /// Count transactions in a solidified block by block number.
+    pub async fn get_transaction_count_by_block_num(&self, block_num: i64) -> Result<u64> {
+        self.inner
+            .get_transaction_count_by_block_num(block_num)
+            .await
+            .map_err(|e| ProviderError::from(e.into()))
+    }
+
+    /// Execute a constant (read-only) contract call against solidified state.
+    pub async fn trigger_constant_contract(
+        &self,
+        params: TriggerSmartContract,
+    ) -> Result<ConstantCallResult> {
+        self.inner
+            .trigger_constant_contract(params)
+            .await
+            .map_err(|e| ProviderError::from(e.into()))
+    }
+
+    /// Estimate the energy a contract call would consume against solidified state.
+    pub async fn estimate_energy(&self, params: TriggerSmartContract) -> Result<i64> {
+        self.inner.estimate_energy(params).await.map_err(|e| ProviderError::from(e.into()))
+    }
+
+    /// Poll until `tx_id` has solidified, regardless of execution result.
+    ///
+    /// Defaults to every 3 s, up to 20 attempts (~60 s).
+    pub async fn wait_for_transaction(
+        &self,
+        tx_id: TxId,
+    ) -> std::result::Result<TransactionInfo, PendingTransactionError> {
+        self.wait_for_transaction_with(tx_id, DEFAULT_POLL_INTERVAL, DEFAULT_POLL_ATTEMPTS).await
+    }
+
+    /// Poll for solidification with a custom interval and attempt count.
+    pub async fn wait_for_transaction_with(
+        &self,
+        tx_id: TxId,
+        interval: Duration,
+        max_attempts: u32,
+    ) -> std::result::Result<TransactionInfo, PendingTransactionError> {
+        for attempt in 0..max_attempts {
+            if attempt > 0 {
+                tokio::time::sleep(interval).await;
+            }
+            match self.get_transaction_info(tx_id).await {
+                Ok(Some(info)) => return Ok(info),
+                Ok(None) => continue,
+                Err(e) => return Err(PendingTransactionError::Transport(e)),
+            }
+        }
+        Err(PendingTransactionError::ConfirmationTimeout)
+    }
+
+    /// Poll until `tx_id` has solidified and its execution succeeded.
+    ///
+    /// Fails with [`PendingTransactionError::Reverted`] (carrying the receipt)
+    /// if the transaction solidified but reverted / ran out of energy.
+    pub async fn wait_for_success(
+        &self,
+        tx_id: TxId,
+    ) -> std::result::Result<TransactionInfo, PendingTransactionError> {
+        self.wait_for_success_with(tx_id, DEFAULT_POLL_INTERVAL, DEFAULT_POLL_ATTEMPTS).await
+    }
+
+    /// Poll for solidified success with a custom interval and attempt count.
+    pub async fn wait_for_success_with(
+        &self,
+        tx_id: TxId,
+        interval: Duration,
+        max_attempts: u32,
+    ) -> std::result::Result<TransactionInfo, PendingTransactionError> {
+        let info = self.wait_for_transaction_with(tx_id, interval, max_attempts).await?;
+        if info.is_success() {
+            Ok(info)
+        } else {
+            Err(PendingTransactionError::Reverted(Box::new(info)))
+        }
+    }
+}
+
+impl SolidityProvider<SolidityGrpcTransport> {
+    /// Connect with the default transport configuration.
+    ///
+    /// Use [`builder`](Self::builder) to customize it.
+    pub async fn connect(uri: impl AsRef<str>) -> Result<Self> {
+        let transport = SolidityGrpcTransport::connect(uri).await.map_err(ProviderError::from)?;
+        Ok(Self::new(transport))
+    }
+
+    /// Start a pre-connect [`SolidityProviderBuilder`].
+    pub fn builder() -> SolidityProviderBuilder {
+        SolidityProviderBuilder::default()
+    }
+}
+
+/// Pre-connect builder for a [`SolidityProvider`] over gRPC.
+#[derive(Clone, Debug, Default)]
+pub struct SolidityProviderBuilder {
+    inner: SolidityGrpcTransportBuilder,
+}
+
+impl SolidityProviderBuilder {
+    /// Override the connect (handshake) timeout.
+    pub fn with_connect_timeout(mut self, timeout: Duration) -> Self {
+        self.inner = self.inner.with_connect_timeout(timeout);
+        self
+    }
+
+    /// Override the per-call request timeout.
+    pub fn with_request_timeout(mut self, timeout: Duration) -> Self {
+        self.inner = self.inner.with_request_timeout(timeout);
+        self
+    }
+
+    /// Override the retry policy.
+    pub fn with_retry(mut self, retry: RetryConfig) -> Self {
+        self.inner = self.inner.with_retry(retry);
+        self
+    }
+
+    /// Add equivalent SolidityNode endpoints for client-side failover.
+    pub fn with_endpoints(mut self, endpoints: Vec<String>) -> Self {
+        self.inner = self.inner.with_endpoints(endpoints);
+        self
+    }
+
+    /// Optionally set the TronGrid API key.
+    pub fn maybe_api_key(mut self, key: Option<impl Into<String>>) -> Self {
+        self.inner = self.inner.maybe_api_key(key);
+        self
+    }
+
+    /// Connect using the accumulated configuration.
+    pub async fn connect(
+        self,
+        uri: impl AsRef<str>,
+    ) -> Result<SolidityProvider<SolidityGrpcTransport>> {
+        let transport = self.inner.connect(uri).await.map_err(ProviderError::from)?;
+        Ok(SolidityProvider::new(transport))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use tronz_primitives::Trx;
+
+    use super::*;
+    use crate::{
+        error::TransportErrorKind,
+        transport::mock::MockSolidityTransport,
+        types::{ContractResult, TxStatus},
+    };
+
+    const NEAR_INSTANT: Duration = Duration::from_millis(1);
+
+    fn info(status: TxStatus) -> TransactionInfo {
+        TransactionInfo {
+            tx_id: TxId::ZERO,
+            block_number: 1,
+            block_timestamp: 1,
+            status,
+            energy_usage: 0,
+            energy_fee: Trx::ZERO,
+            net_usage: 0,
+            net_fee: Trx::ZERO,
+            contract_result: ContractResult::Default,
+            contract_address: None,
+            logs: vec![],
+            revert_reason: None,
+        }
+    }
+
+    fn provider(mock: MockSolidityTransport) -> SolidityProvider<MockSolidityTransport> {
+        SolidityProvider::new(mock)
+    }
+
+    #[tokio::test]
+    async fn wait_for_transaction_polls_past_none_then_returns_receipt() {
+        let mock = MockSolidityTransport::new();
+        mock.push_ok::<Option<TransactionInfo>>("get_transaction_info", None).push_ok::<Option<
+            TransactionInfo,
+        >>(
+            "get_transaction_info",
+            Some(info(TxStatus::Failed)),
+        );
+
+        let receipt =
+            provider(mock).wait_for_transaction_with(TxId::ZERO, NEAR_INSTANT, 5).await.unwrap();
+        assert!(!receipt.is_success());
+    }
+
+    #[tokio::test]
+    async fn wait_for_success_rejects_reverted_and_preserves_receipt() {
+        let mock = MockSolidityTransport::new();
+        mock.push_ok::<Option<TransactionInfo>>(
+            "get_transaction_info",
+            Some(info(TxStatus::Failed)),
+        );
+
+        let err =
+            provider(mock).wait_for_success_with(TxId::ZERO, NEAR_INSTANT, 5).await.unwrap_err();
+        match err {
+            PendingTransactionError::Reverted(receipt) => assert!(!receipt.is_success()),
+            other => panic!("expected Reverted, got {other:?}"),
+        }
+    }
+
+    #[tokio::test]
+    async fn wait_for_success_returns_successful_receipt() {
+        let mock = MockSolidityTransport::new();
+        mock.push_ok::<Option<TransactionInfo>>(
+            "get_transaction_info",
+            Some(info(TxStatus::Success)),
+        );
+
+        let receipt =
+            provider(mock).wait_for_success_with(TxId::ZERO, NEAR_INSTANT, 5).await.unwrap();
+        assert!(receipt.is_success());
+    }
+
+    #[tokio::test]
+    async fn wait_for_transaction_times_out_when_never_solidified() {
+        let mock = MockSolidityTransport::new();
+        for _ in 0..3 {
+            mock.push_ok::<Option<TransactionInfo>>("get_transaction_info", None);
+        }
+
+        let err = provider(mock)
+            .wait_for_transaction_with(TxId::ZERO, NEAR_INSTANT, 3)
+            .await
+            .unwrap_err();
+        assert!(matches!(err, PendingTransactionError::ConfirmationTimeout));
+    }
+
+    #[tokio::test]
+    async fn zero_attempts_times_out_without_polling() {
+        let err = provider(MockSolidityTransport::new())
+            .wait_for_transaction_with(TxId::ZERO, NEAR_INSTANT, 0)
+            .await
+            .unwrap_err();
+        assert!(matches!(err, PendingTransactionError::ConfirmationTimeout));
+    }
+
+    #[tokio::test]
+    async fn transport_error_propagates() {
+        let mock = MockSolidityTransport::new();
+        mock.push_err::<Option<TransactionInfo>>(
+            "get_transaction_info",
+            TransportErrorKind::Malformed("boom".to_owned()),
+        );
+
+        let err = provider(mock)
+            .wait_for_transaction_with(TxId::ZERO, NEAR_INSTANT, 5)
+            .await
+            .unwrap_err();
+        assert!(matches!(err, PendingTransactionError::Transport(_)));
+    }
+}
