+++ crates/provider/tests/support/Cargo.toml
@@ -1,0 +1,23 @@
+# Standalone test-support crate: a deterministic local `protocol.WalletSolidity`
+# gRPC server for `tronz-provider`'s SolidityNode transport tests.
+#
+# Its own `[workspace]` table detaches it from the parent workspace so its
+# test-only protobuf/server generation never touches the production build.
+[package]
+name = "tronz-provider-test-support"
+version = "0.0.0"
+edition = "2024"
+publish = false
+
+[lib]
+
+[dependencies]
+tonic        = { version = "0.14", default-features = false, features = ["transport", "codegen", "server", "router"] }
+tonic-prost  = { version = "0.14" }
+prost        = { version = "0.14" }
+prost-types  = { version = "0.14" }
+tokio        = { version = "1", features = ["rt-multi-thread", "macros", "net", "sync", "time"] }
+tokio-stream = { version = "0.1", features = ["net"] }
+
+[build-dependencies]
+tonic-prost-build = { version = "0.14" }
